moved: subcommand to seperate files
